# app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server We Will Guide You server.",
        "version": "1.0.0",
        "title": "Swagger We Will Guide You",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/manage/guides": {
            "post": {
                "tags": [
                    "Guide"
                ],
                "summary": "Add a new Guide",
                "description": "Adds a new Guide to the Specific Subject.",
                "operationId": "addGuide",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Guide model that needs to be added",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/AddGuideAdminApiModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Guide added successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailGuideAdminApiModel"
                        }
                    },
                    "405": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "get": {
                "tags": [
                    "Guide"
                ],
                "summary": "Get a list of Guide",
                "description": "It's an Api for fetching All The Guides.",
                "operationId": "listGuide",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "query",
                        "name": "offset",
                        "description": "Number of entities to skip.",
                        "required": false,
                        "default": 0,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "Number of entities to get.",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "query",
                        "description": "Query for searching and sorting.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide list",
                        "schema": {
                            "$ref": "#/definitions/GuideListAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
...
        "/manage/references": {
            "post": {
                "tags": [
                    "Reference"
                ],
                "summary": "Add a new Reference",
                "description": "Adds a new Reference.",
                "operationId": "addReference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Reference model that needs to be added",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/AddReferenceAdminApiModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Reference added successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailReferenceAdminApiModel"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "get": {
                "tags": [
                    "Reference"
                ],
                "summary": "Get a list of Reference",
                "description": "It's an Api for fetching All The References.",
                "operationId": "listReference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "parentId",
                    "description": "Id of parent.",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "References found",
                        "schema": {
                            "$ref": "#/definitions/ReferenceListAdminApiModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "definitions": {
...
        "VoidResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "ResponseMessage": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "number"
                },
                "messageId": {
                    "type": "number"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "error",
                        "warning",
                        "info"
                    ]
                }
            }
        },
...
        "AddGuideAdminApiModel": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "priorityOrder": {
                    "type": "number"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subjectId": {
                    "type": "string"
                }
            }
        },
...
        "DetailGuideAdminApiModel": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "text": {
                            "type": "string"
                        },
                        "priorityOrder": {
                            "type": "number"
                        },
                        "isEnabled": {
                            "type": "boolean"
                        },
                        "lastModifiedDateTime": {
                            "type": "object"
                        },
                        "tags": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "author": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "firstName": {
                                    "type": "string"
                                },
                                "lastName": {
                                    "type": "string"
                                }
                            }
                        },
                        "subject": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "GuideListAdminApiModel": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "result": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "isEnabled": {
                                "type": "boolean"
                            },
                            "lastModifiedDateTime": {
                                "type": "object"
                            },
                            "author": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    }
                                }
                            },
                            "subject": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "totalCount": {
                    "type": "number"
                }
            }
        },
        "AddReferenceAdminApiModel": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "priorityOrder": {
                    "type": "number"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
...
        "DetailReferenceAdminApiModel": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "text": {
                            "type": "string"
                        },
                        "priorityOrder": {
                            "type": "number"
                        },
                        "isEnabled": {
                            "type": "boolean"
                        },
                        "lastModifiedDateTime": {
                            "type": "object"
                        },
                        "parent": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            }
                        },
                        "lastModifierUser": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "ReferenceListAdminApiModel": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "result": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "isEnabled": {
                                "type": "boolean"
                            }
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "totalCount": {
                    "type": "number"
                }
            }
        },
...
    },
...
}